requirements.txt
eventlet==0.33.0
django-cors-headers==3.11.0
Django==4.0.3
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.3
gunicorn==20.0.4